+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### v1.1.0
   - Throw an error when a relative path is passed to the `render` function. Only absolute path are accepted.
+  - It is possible retry to render a report with the following 2 configurations:
+    - `retriesOnError` is the number of retries before returning the error. Default value: 1.
+    - `retriesIntervalOnError` is the interval of time before retrying. Default value: 0ms.
+
+    To change the default value, the `setOptions` method have to be used: `sdk.setOptions({ retriesOnError: 5, retriesIntervalOnError: 2000 });`
 
 ### v1.0.0
   - Release July 3rd, 2020